# weight/docs/swagger.yaml
definitions:
  structs.AddWeight:
    properties:
      date:
        example: "2021-07-24"
        type: string
      max:
        example: 55
        type: number
      min:
        example: 50
        type: number
      source:
        example: Postman
        type: string
      userId:
        example: Postman
        type: string
      userName:
        example: Postman
        type: string
    required:
    - date
    - max
    - min
    - userId
    type: object
  structs.ResponseAddWeight:
    properties:
      data:
        $ref: '#/definitions/structs.TxWeightRecord'
      message:
        example: Add weight successfully
        type: string
      status:
        example: SUCCESS
        type: string
    type: object
  structs.ResponseError400:
    properties:
      data:
        type: object
      message:
        example: 'Key: ''GetRestaurantTrx.RestName'' Error:Field validation for ''RestName''
          failed on the ''required'' tag'
        type: string
      status:
        example: FAILED
        type: string
    type: object
  structs.ResponseError500:
    properties:
      data:
        type: object
      message:
        example: Restaurant not found
        type: string
      status:
        example: FAILED
        type: string
    type: object
  structs.ResponseGetWeight:
    properties:
      data:
        items:
          $ref: '#/definitions/structs.TxWeightRecord'
        type: array
      message:
        example: Get weight successfully
        type: string
      status:
        example: SUCCESS
        type: string
    type: object
  structs.ResponseUpdateWeight:
    properties:
      data:
        $ref: '#/definitions/structs.TxWeightRecord'
      message:
        example: Update weight successfully
        type: string
      status:
        example: SUCCESS
        type: string
    type: object
  structs.TxWeightRecord:
    properties:
      created_by:
        example: DBeaver
        type: string
      created_date:
        example: "2021-07-20 09:34:12"
        type: string
      created_from:
        example: DBeaver
        type: string
      created_name:
        example: DBeaver
        type: string
      date:
        example: "2021-07-24"
        type: string
      max:
        example: "55"
        type: string
      min:
        example: "50"
        type: string
      modified_by:
        example: DBeaver
        type: string
      modified_date:
        example: "2021-07-20 09:34:12"
        type: string
      modified_from:
        example: DBeaver
        type: string
      modified_name:
        example: DBeaver
        type: string
      status_id:
        example: "1"
        type: string
      weight_record_id:
        example: ddccc2ad-4ebe-4b59-92b1-8417d5e5a0cf
        type: string
    type: object
  structs.UpdateWeight:
    properties:
      date:
        example: "2021-07-24"
        type: string
      max:
        example: 55
        type: number
      min:
        example: 50
        type: number
      source:
        example: Postman
        type: string
      userId:
        example: Postman
        type: string
      userName:
        example: Postman
        type: string
    required:
    - userId
    type: object
info:
  contact: {}
  description: Example API for Weight
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Weight
  version: "1.0"
paths:
  /api/v1/weights:
    get:
      consumes:
      - application/json
      description: Add Weight API
      parameters:
      - description: 'Example: 2021-04-01'
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ResponseGetWeight'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ResponseError400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ResponseError500'
      summary: API for Add Weight
    post:
      consumes:
      - application/json
      description: Add Weight API
      parameters:
      - description: Body of Weight
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/structs.AddWeight'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ResponseAddWeight'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ResponseError400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ResponseError500'
      summary: API for Add Weight
  /api/v1/weights/{weightRecordId}:
    put:
      consumes:
      - application/json
      description: Update Weight API
      parameters:
      - description: 'Example: 419b68d8-fe6f-4025-b8e5-dcb572f74b27'
        in: path
        name: weightRecordId
        required: true
        type: string
      - description: Body of Weight
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateWeight'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ResponseUpdateWeight'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ResponseError400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ResponseError500'
      summary: API for Update Weight
swagger: "2.0"